backend/requirements.txt
fastapi[standard]
pymongo
python-dotenv
pydantic[email]
python-multipart
aiofiles
boto3
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
google-auth
httpx
qrcode[pil]
Pillow
